Move Multus pod annotation generation logic
As part of the effort to consolidate sig-network's
code under sig-network's ownership,
move `GenerateMultusCNIAnnotation` and the functions
it depends on under the `multus` package.

Also rename them due to the fact they are now
invoked with the `multus` package prefix.

This is done in order to enable moving additional
logic under pkg/network.
