+++ pkg/network/multus/BUILD.bazel
@@ -9,7 +9,10 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/network/multus",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/network/namescheme:go_default_library",
+        "//pkg/network/netbinding:go_default_library",
         "//pkg/network/vmispec:go_default_library",
+        "//pkg/virt-config:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/precond:go_default_library",
         "//vendor/github.com/k8snetworkplumbingwg/network-attachment-definition-client/pkg/apis/k8s.cni.cncf.io/v1:go_default_library",
@@ -24,6 +27,8 @@ go_test(
     ],
     deps = [
         ":go_default_library",
+        "//pkg/testutils:go_default_library",
+        "//pkg/virt-config:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
