+++ pkg/virt-controller/network/BUILD.bazel
@@ -17,9 +17,7 @@ go_library(
         "//pkg/network/downwardapi:go_default_library",
         "//pkg/network/multus:go_default_library",
         "//pkg/network/namescheme:go_default_library",
-        "//pkg/network/netbinding:go_default_library",
         "//pkg/network/vmispec:go_default_library",
-        "//pkg/virt-config:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/generated/kubevirt/clientset/versioned:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
@@ -36,7 +34,6 @@ go_test(
     name = "go_default_test",
     srcs = [
         "hotplug_test.go",
-        "multus_annotations_test.go",
         "network_suite_test.go",
         "pod_annotations_test.go",
         "render_test.go",
@@ -47,8 +44,6 @@ go_test(
         "//pkg/libvmi:go_default_library",
         "//pkg/network/downwardapi:go_default_library",
         "//pkg/network/vmispec:go_default_library",
-        "//pkg/testutils:go_default_library",
-        "//pkg/virt-config:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/generated/kubevirt/clientset/versioned/fake:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
