+++ pkg/virt-controller/watch/BUILD.bazel
@@ -28,6 +28,7 @@ go_library(
         "//pkg/monitoring/metrics/virt-controller:go_default_library",
         "//pkg/monitoring/profiler:go_default_library",
         "//pkg/network/admitter:go_default_library",
+        "//pkg/network/multus:go_default_library",
         "//pkg/network/namescheme:go_default_library",
         "//pkg/network/netbinding:go_default_library",
         "//pkg/network/vmispec:go_default_library",
@@ -127,6 +128,7 @@ go_test(
         "//pkg/instancetype:go_default_library",
         "//pkg/monitoring/metrics/virt-controller:go_default_library",
         "//pkg/network/downwardapi:go_default_library",
+        "//pkg/network/multus:go_default_library",
         "//pkg/network/vmispec:go_default_library",
         "//pkg/pointer:go_default_library",
         "//pkg/rest:go_default_library",
@@ -136,7 +138,6 @@ go_test(
         "//pkg/storage/types:go_default_library",
         "//pkg/testutils:go_default_library",
         "//pkg/virt-config:go_default_library",
-        "//pkg/virt-controller/network:go_default_library",
         "//pkg/virt-controller/services:go_default_library",
         "//pkg/virt-controller/watch/clone:go_default_library",
         "//pkg/virt-controller/watch/descheduler:go_default_library",
